accounted for overflow in manhattanDistance

+++ src/main/scala/io/github/reggert/scalademo/scalaexamples/package.scala
@@ -2,8 +2,8 @@ package io.github.reggert.scalademo
 
 package object scalaexamples {
   implicit final class PointOps(val self : Point) extends AnyVal {
-    def manhattanDistance(other : Point) : Int = {
-      self.x - other.x + self.y - other.y + self.z - other.z
+    def manhattanDistance(other : Point) : Long = {
+      self.x.toLong - other.x.toLong + self.y.toLong - other.y.toLong + self.z.toLong - other.z.toLong
     }
   }
 }
